requirements.txt
dataset
nose
Flask-Bootstrap
git+https://github.com/maxcountryman/flask-uploads.git@f66d7dc